[FEAT] 가게 리뷰 목록 조회 API #14

+++ src/main/java/umc/spring/web/dto/ReviewRequestDTO.java
@@ -8,7 +8,7 @@ public class ReviewRequestDTO {
 
     @Getter
     public static class WriteDto{
-        int score;
+        Float score;
         String contents;
         Long memberId;
         Long storeId;
